Bouge les scripts de generation de tiles dans le dossier tiles pour plus de lisibilité
